KVService proxy attr access to configured backend
This way we can use all the methods exposed by the backend without
having to mirror each one in our API. Possible for this to cause trouble
if we try to add something like memcache in the future and we're using
some Redis specific methods... could end up with a compatibility layer.
No point "over-engineering" now.